+++ LearningCSharp7/Source/Learn.CSharp7.Common/Utilis/ExtensionMethods.cs
@@ -33,7 +33,7 @@ namespace Learn.CSharp7.Common.Utilis
             salary = student.Salary;
         }
 
-        public static void Deconstruct(this Subject subject, out string name, out decimal marks)
+        public static void Deconstruct(this Subject subject, out string name, out float marks)
         {
             name = subject.Name;
             marks = subject.Marks;
